LPD-57471 Update unit tests for SpaceMembersWithList component
